[#38] refactor: acronym utils renamed to string utils, string slicer implemented
